just adds a button to the storage panel in settings that takes you to whatever directory you set it as previously.
I also decided to work on the font size button on this branch to so yeah theres that. It is in ai models underneath confidence threshold. I've also added a lil preview to kinda show what the tracking log will look like when you choose your font size so lemme know if that's worth keeping. I made it to where the colors you change for the log message will also appear on the log message preview btw.